app: scss: fixed container being too short in Safari browser
There was a problem viewing the site in Safari on Windows where the
container wasn't a minimum of 100% of the viewport's height.  This was
fixed by first ensuring that the html element was a minimum of 100% of the
viewport's height and then making the body element match those dimensions
and then finally having .container be at minimum of 100% height.