app: made the sortability of columns more visible to users using underlining and colour
This was based on a comment from Stephanos:
"
What I would recommend thought is to highlight (change
color) the column title that is sorted, by default is NAME and perhaps
underline the other 2 columns indicating that those are links as well. As
it is right now, users might not understand that these are actually links
that can sort the columns.
"

Related to #135