added nullable external_id field to location_duplicate table and improved filtering of duplicate locations
The external_id field is to maintain more specific information on location
duplicates so we can more confidently rule that a location from an
external source is a duplicate and shouldn't be added by matching the
other data source's id field.

This commit also improves filtering of duplicate locations by making use
of the location_duplicate data.

Related to #253