removed needless prototype from Hacking Health 2016
The prototype from Hacking Health 2016 wasn't serving a purpose any
longer.

The files were reviewed for potential use and 300 locations were taken
from one of its sql files in connection with #54.

A web_demo created by Bruce in #33 are being removed because the Google
maps related to it have been implemented and there's no need to ever look
at the web_demo.

The requirements for the application have been thoroughly discussed for
almost a year so the demo quite surely says nothing about the requirements
that we don't know and have documented in a better way.

Having these files linger around is now nothing but a source of confusion
for new team members.  We're not using NodeJS but Laravel as a web
framework.  Our application is in the app directory and not the web directory...