app: profile: improved layout for viewports narrower than 400 pixels
This included reducing whitespace in the accordion categories to make
room for question text.

display flex was used to fix wrapped question text to a new line instead
of lining up vertically with the checkbox.

Some heading font sizes were reduced because text like 'My Accessibility
Requirements' was wrapping on narrow viewports.  The full name was likely
to wrap if the name was long enough so that font was reduced too.