app scss: reduced whitespace for narrow viewports on FAQ page and improved layouts for location search
The 25 pixel padding looked excessive for viewports narrower than 300
pixels so I reduced it to 10 pixels when the viewport is that narrow.

This also makes the location search look better on narrow viewports.  This
fixes a problem where 'Accessibility Rating' wrapped while testing in an
Samsung S3 Android tablet.  This also reduces padding in most pages for
narrow viewports to free up space for important content.  The home page
needed some SCSS tweaks to maintain its original lack of padding on narrow
viewports.