app location rating: fixed bug where location rating requests could still be in progress before navigation completes
I noticed that if lots of AJAX requests were fired rapidly, navigating
away from the page would cause some of the AJAX requests to fail and
not all changes would be get loaded when visiting the page again.
The changes in this commit replace all
anchor(a) tag navigation with a call to a function that waits until the
AJAX calls complete.  A similar technique is used to delay submission of
the 'Submit' form.  It also uses beforeunload event to block requests to
reload the page or close the browser before things save.

This isn't a perfect fix.  The beforeUnload event isn't
supported by all browsers.  Closing a browser that doesn't support
beforeUnload would still leave some data unsaved.  Reloading the rating
page when the browser doesn't support onbeforeunload would also cause
unsaved changes to be lost.  This is just as good as I could do for now.