Question Category sorted on various pages
This includes location rating, report, and user profile.

This also includes some navigation updates so that 'Next' and 'Submit'
buttons in the rating feature stay consistent with the new order of
question categories.

Some inconsistent use of tabs and spaces was fixed with indentation for
LocationRatingController.php.

Done as part of #244