app: added location tag icons when they fit better than full text
This included a migration to add the icon_selector field to the
location_tag table.

This also included some changes to the json_to_db tool to merge the new
icon information into deployed databases.

This also included some JavaScript and SCSS changes to choose the icon
layout when the viewport width and number of location tags suggest that
the icons will fit better.

Related to #430