app: scss: improved responsiveness of location map report
The main thing this commit does is stack the question category box above
the map for viewports narrower than 992 pixels.

This might fix a bug with Universal and Personal links wrapping on
Yashankit's iphone.

Related to feedback from Yashankit on #430 where he found the Universal
and Personal buttons wrapping to new lines.