app: removed Remember Me checkbox because it isn't working
Until someone completes #383 with a working Remember Me feature or has
some other reasonable purpose behind it, the checkbox is just confusing.
The checkbox was part of an incomplete feature and incomplete features
tend to confuse people who test the application.  It also requires effort
to communicate that we know it is incomplete.  Removing incomplete
features seems better for now.

Related to #383.