app: limited location search radius to maximum 500km and supported center move in map
Setting the circle to editable in a previous pull request for #373 was
intended just to let the user adjust the circle's radius but it
unfortunately also let the user move the center without doing anything
with it.  One of the changes in this commit saves the new center location
along with an appropriate address.

Related to #373