app: added a feature to delete users
This feature is intended to help Blandine demo the application to people.
She was having trouble signing into the application and wants to show the
signup process.  The signup process doesn't work for email addresses
already signed up, though.

This feature is restricted to just Blandine's email addresses for now to
prevent loss of important answer data and abuse.