app: made it so signing in with facebook or google third party authentication updates the email confirmation date
There was a problem with the following scenario:
1. Sign up a new account using third party authentication. (no email
confirmation email sent and no confirmation date set)
2. Perform password recovery.  (No password to recover)
3. Try to sign in.  It says to check email to confirm it but there is
none.  It is like the app thought the email must have been sent when the
signup happened despite the sign up working through third-party
authentication.

This fixes it by trusting that third party authentication is indirectly
verifying that the email address is tied to the same person.