app: AnswerRepository.php: worked around cases where the rating cache is stale
If new questions are added but the cache isn't refreshed, it is possible
for the ratings_cache field to be not null yet not have one of the
questions.  This commit prevents the user from seeing an error message
in this situation.  The case should be extremely rare.  Only in the
middle of a deployment should this ideally happen since the cache should
be refreshed shortly after any new questions are added or modified.

Related to #555