app: added soft deletes for user_answer table
Also, made use of the new field to simplify a rating calculation query.

Also, adjusted some of the AnswerRepository to maintain the new deleted_at field in its intended state. (NULL for only the latest answer by a given user for a given location and question).

I looked for but couldn't get the Eloquent delete method and queries to recognize deleted_at for the soft-delete field it was intended to be.  Instead, resorted to explicitly using update statements, whereNull('deleted_at').

Related to #615