app: updated phpunit from 5.6 to 7.1 and updated Laravel from 5.3 to 5.4
This required updating virtually every PHPUnit test to stop using seeStatusCode and instead assert that the result of getStatusCode matched the value.

This also involved updating several other dependencies in composer.json and the lock file.

Some minor fixes were made in response to the merge of pull request #620.  These changes were in the SigninTest.php file.  When saving profile information, the server redirected to the home page.  The tests were failing for this because they weren't adjusted appropriately in #620.