app: replaced Eloquent with Model for several model classes
This was done as part of getting Hardik's changes for #615 in master.

Eloquent and Model are essentially the same thing so that shouldn't change much.  More detail is at:
https://laracasts.com/discuss/channels/eloquent/models-vs-classes

A few other changes were made in the models such as using hasMany in 1-N relationships.  The belongsToMany is mainly for M-N relationships.