importers: json_to_db.py: optimized to drastically speed up the json_to_db script
Using a set is way faster than a list because it reduces an "in" operation from O(n) to O(1) which makes a big difference when there are over 20000 elements in a list.

This speed up reduced the time to sync from roughly 5 minutes down to 5 seconds when I tested on my laptop.